# openapi/spec.yaml
openapi: 3.1.0
info:
  title: OAuth Implementation Server
  description: |
    A comprehensive OAuth 2.0 / OpenID Connect implementation server for testing all OAuth flows and standards.

    Supports: Authorization Code + PKCE, Client Credentials, Device Code (RFC 8628),
    CIBA (Client-Initiated Backchannel Authentication), PAR (RFC 9126), DPoP (RFC 9449),
    Token Introspection (RFC 7662), Token Revocation (RFC 7009), Dynamic Client Registration (RFC 7591),
    Resource Indicators (RFC 8707), OIDC Discovery, JWKS, UserInfo, TOTP-based MFA,
    Consent Management, Custom Scopes, and login_hint/prompt/max_age parameters.
  version: 1.1.0
  contact:
    name: API Support
    url: https://github.com/bravo68web/oauth-impl
servers:
  - url: https://localhost:8080
    description: Local development server
...
paths:
...
  /oauth/logout:
    get:
      tags: [OIDC]
      summary: RP-Initiated Logout
      description: Ends the browser session. A valid id_token_hint logs out without a prompt. Otherwise the response is a confirmation page.
      parameters:
        - in: query
          name: id_token_hint
          schema: { type: string }
        - in: query
          name: client_id
          schema: { type: string }
        - in: query
          name: post_logout_redirect_uri
          schema: { type: string }
        - in: query
          name: state
          schema: { type: string }
      responses:
        "200":
          description: Logged out, or a confirmation page
        "302":
          description: Redirect to a registered post_logout_redirect_uri
    post:
      tags: [OIDC]
      summary: Confirm RP-Initiated Logout
      responses:
        "200":
          description: Logged out
...
  /oauth/device:
    post:
      tags: [Device]
      summary: Device authorization (RFC 8628)
      description: |
        Initiates the Device Authorization Grant for input-constrained devices (smart TVs, IoT, CLI tools).

        Returns a `device_code`, `user_code`, and `verification_uri`. The user visits the verification URI
        on another device and enters the user code. The client polls `/oauth/token` with the device_code
        until the user completes authorization.

        See [RFC 8628](https://datatracker.ietf.org/doc/html/rfc8628) for details.
      operationId: deviceAuthorization
      security:
        - basicAuth: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                scope:
                  type: string
                  description: Requested scopes
                client_id:
                  type: string
                  description: Client ID (alternative to HTTP Basic auth)
                client_secret:
                  type: string
                  description: Client secret (alternative to HTTP Basic auth)
            example:
              scope: "openid profile"
      responses:
        "200":
          description: Device authorization initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceAuthorizationResponse"
              example:
                device_code: "GmRhmhcxhwAzkoEqiMEg_DnyEysNkuNhszIySk9eS"
                user_code: "WDJB-MJHT"
                verification_uri: "http://localhost:8080/device"
                verification_uri_complete: "http://localhost:8080/device?user_code=WDJB-MJHT"
                expires_in: 1800
                interval: 5
        "400":
          description: Client not authorized for device_code grant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthError"
              example:
                error: unauthorized_client
                error_description: "Client not authorized for device_code grant"
...
  /device:
    get:
      tags: [Device]
      summary: Device verification page
      description: |
        HTML form where the user enters the user_code displayed on their device.
        Optionally accepts `user_code` as a query parameter to pre-fill the form.
      operationId: deviceVerificationPage
      parameters:
        - name: user_code
          in: query
          schema:
            type: string
          description: Pre-filled user code from the device
          example: "WDJB-MJHT"
      responses:
        "200":
          description: HTML verification form
          content:
            text/html:
              example: "<html><body><form>Enter code: <input name='user_code'/></form></body></html>"
    post:
      tags: [Device]
      summary: Submit device verification
      description: |
        Processes the device verification form. If the user_code is valid and not expired,
        marks the device code as approved.
      operationId: deviceVerificationSubmit
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [user_code]
              properties:
                user_code:
                  type: string
                  description: The code displayed on the device
                  example: "WDJB-MJHT"
                user_id:
                  type: string
                  description: User ID for approval
      responses:
        "200":
          description: Verification approved (HTML confirmation page)
          content:
            text/html:
              example: "<html><body><h1>Authorization Approved</h1><p>You can return to your device.</p></body></html>"

  # ──────────────────────────────────────────────
  # CIBA Polling (mobile-facing)
  # ──────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Opaque OAuth 2.0 access token. Management routes require a client-credentials token whose scope contains management.
    basicAuth:
      type: http
      scheme: basic
      description: HTTP Basic authentication with client_id:client_secret
  schemas:
...
    DeviceAuthorizationResponse:
      type: object
      description: Device authorization response (RFC 8628)
      example:
        device_code: "GmRhmhcxhwAzkoEqiMEg_DnyEysNkuNhszIySk9eS"
        user_code: "WDJB-MJHT"
        verification_uri: "http://localhost:8080/device"
        verification_uri_complete: "http://localhost:8080/device?user_code=WDJB-MJHT"
        expires_in: 1800
        interval: 5
      properties:
        device_code:
          type: string
          description: Device verification code (used for polling, not shown to user)
          example: "GmRhmhcxhwAzkoEqiMEg_DnyEysNkuNhszIySk9eS"
        user_code:
          type: string
          description: "User code (shown to user, e.g. WDJB-MJHT)"
          example: "WDJB-MJHT"
        verification_uri:
          type: string
          description: URL where user enters the user_code
          example: "http://localhost:8080/device"
        verification_uri_complete:
          type: string
          description: URL with user_code pre-filled (can show as QR code)
          example: "http://localhost:8080/device?user_code=WDJB-MJHT"
        expires_in:
          type: integer
          description: Lifetime of device_code and user_code in seconds
          example: 1800
        interval:
          type: integer
          description: Minimum polling interval in seconds
          example: 5

    # ── PAR ──
...
    OAuthError:
      type: object
      required: [error]
      description: OAuth 2.0 error response
      example:
        error: invalid_request
        error_description: "client_id is required"
      properties:
        error:
          type: string
          description: Error code
          enum:
            - invalid_request
            - invalid_client
            - invalid_grant
            - unauthorized_client
            - unsupported_grant_type
            - invalid_scope
            - access_denied
            - server_error
            - invalid_token
            - invalid_dpop_proof
            - expired_token
            - authorization_pending
            - slow_down
            - invalid_redirect_uri
          example: invalid_request
        error_description:
          type: string
          description: Human-readable error description
          example: "client_id is required"
        state:
          type: string
          description: State parameter from the original request (echoed back)
          example: "abc123xyz"